[프로젝트 삭제] : 소유주가 아닐 시 삭제 불가 조건 추가

+++ src/main/java/com/lms/api/admin/project/ProjectService.java
@@ -124,6 +124,9 @@ public class ProjectService {
         ProjectMemberEntity projectMemberEntity = projectMemberRepository.findByProjectEntity_IdAndUserEntity_IdAndProjectRole(projectId, userId, ProjectRole.OWNER)
                 .orElseThrow(() -> new ApiException(ApiErrorCode.ACCESS_DENIED));
 
+        if(projectMemberEntity.getProjectRole() != ProjectRole.OWNER){
+            throw new ApiException(ApiErrorCode.ACCESS_DENIED);
+        }
         projectRepository.delete(projectEntity);
     }
